🐛 portfpolioとcontactのタイポを修正しフォルダをtestからmainに移動
