--- components/Organisms/PFCardCom2.tsx
@@ -1,21 +1,0 @@
-import { FC } from "react"
-import { useQueryPoFo } from "hooks/useQueryPost"
-import { PFCard } from "components/Molecules/PFCard"
-
-export const PFCardCom: FC = () => {
-  const { data } = useQueryPoFo()
-
-  return (
-    <>
-      {data?.contents.map((item) => (
-        <PFCard
-          key={item.id}
-          url={item.url}
-          title={item.title}
-          body={item.body}
-          publishedAt={item.publishedAt}
-        />
-      ))}
-    </>
-  )
-}
