+++ pages/contact.tsx
@@ -1,49 +1,69 @@
-import Head from "next/head"
 import type { NextPage } from "next"
-import { Layout } from "components/FixedElement/Layout"
-
 import { Center, Stack, Textarea, TextInput, Button } from "@mantine/core"
 import { useForm } from "@mantine/form"
 
+import { Layout } from "components/FixedElement/Layout"
+
+type FormValues = {
+  email: string
+  name: string
+  message: string
+}
+
 const Contact: NextPage = () => {
-  const form = useForm({
+  const form = useForm<FormValues>({
     initialValues: {
       email: "",
       name: "",
       message: "",
     },
     validate: {
-      email: (value) => (/^\S+@\S+$/.test(value) ? null : "Invalid email"),
+      email: (value) => (/^\S+@\S+$/.test(value) ? null : "入力してください"),
     },
   })
+
+  const handleSubmit = async (values: FormValues) => {
+    await fetch("/api/contacts", {
+      method: "POST",
+      body: JSON.stringify({
+        email: values.email,
+        name: values.name,
+        message: values.message,
+      }),
+    })
+    form.reset()
+  }
+
   return (
     <Layout title="Contact">
-      <Head>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <form onSubmit={form.onSubmit((values) => console.log(values))}>
+      <form
+        onSubmit={form.onSubmit(handleSubmit)}
+        style={{ position: "relative" }}
+      >
+        {/* <LoadingOverlay
+          visible={processing.current}
+        /> */}
         <Stack spacing="xl">
           <TextInput
             required
-            label="Email"
+            label="メールアドレス"
             placeholder="your@email.com"
             {...form.getInputProps("email")}
           />
           <TextInput
             required
-            label="Name"
-            placeholder="Taro Yamada"
+            label="お名前"
+            placeholder="島田　光一郎"
             {...form.getInputProps("name")}
           />
           <Textarea
             required
-            label="Your message"
-            placeholder="I want to order your goods"
+            label="お問い合わせ内容"
+            placeholder=""
             {...form.getInputProps("message")}
           />
           <Center>
-            <Button type="submit">Send Message</Button>
+            <Button type="submit">送信</Button>
           </Center>
         </Stack>
       </form>
